% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clear.R
\name{clear}
\alias{clear}
\title{Clear Console in RStudio}
\usage{
clear()
}
\description{
This function clears the console equivalent to \code{Ctrl + L} in RStudio on
Windows, Mac, UNIX, or Linux operating system.
}
\examples{
\dontrun{

# Clear console
clear()
}
}
\seealso{
\code{\link{restart}}, \code{\link{setsource}}
}
\author{
Takuya Yanagida
}
