% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item.distract.R
\name{item.distract}
\alias{item.distract}
\title{Distractor Analysis for Multiple-Choice Items}
\usage{
item.distract(data, ..., key, exclude = NULL, na.keep = FALSE, correct = TRUE,
              method = c("pbiser", "biser"), ml = FALSE,
              print = c("all", "nUQ", "key", "n", "perc", "r"),
              color = "default", style = c("regular", "bold", "italic"),
              na.omit = FALSE, digits = 1, r.digits = 3, as.na = NULL,
              write = NULL, append = TRUE, check = TRUE, output = TRUE)
}
\arguments{
\item{data}{a data frame. Note that missing values should be represented
with \code{NA} and not with a valid code. The argument
\code{as.na} can be used to convert values to \code{NA}
before conducting the analysis.}

\item{...}{an expression indicating the variable names in \code{data}
e.g., \code{item.distract(dat, x1, x2, x3)}. Note that the
operators \code{+}, \code{-}, \code{~}, \code{:}, \code{::},
and \code{!} can also be used to select variables, see
'Details' in the \code{\link{df.subset}} function.}

\item{key}{a numeric or character vector of the answer key with a length
matching the number of variables specified in the argument
\code{data} or selected using the argument \code{...}.}

\item{exclude}{a character vector indicating items to be excluded from
the analysis.}

\item{na.keep}{logical: if \code{FALSE} (default), missing values are coded
as \code{0} when scoring items (i.e., item not solved) and
creating groups (i.e., answer option not selected).}

\item{correct}{logical: if \code{TRUE} (default), the corrected distractor-total
correlation is computed.}

\item{method}{a character string indicating which correlation coefficient
is used for the item-total correlation, i.e., \code{"pbiser"}
for the point-biserial correlation coefficient when data
are dichotomous, \code{"biser"} for the biserial correlation
coefficient.}

\item{ml}{logical: if \code{FALSE} (default), a two-step approximation
is used to compute the biserial correlation coefficient
(\code{method = "biser"}), while the maximum-likelihood (ML)
estimate is computed if \code{TRUE}. Note that ML estimation
is computationally expensive, i.e., takes a lot of time.}

\item{print}{a character vector indicating which results to print, i.e.,
\code{"all"} for all results, \code{"nUQ"} for the number
of unique elements after omitting missing values, \code{"key"}
for the answer key, \code{"n"} for the distractor selection
frequency, \code{"perc"} for the distractor selection percentage,
and \code{"r"} for the (corrected) distractor-total correlation.}

\item{color}{a character string indicating the text color for highlighting
the attractor frequency, percentage, and attractor-total
correlation, , i.e., \code{"default"} (default)
for the default text color without color coding and various
text colors for highlighting like \code{"red"}, \code{"b.red"},
\code{"green"}, \code{"b.green"}, \code{"blue"}, or \code{"b.blue"},
see the help page of the \code{\link{chr.color}} function.
Note that this option is not supported when using R Markdown and when writing the
output into a text file (\code{.txt}).}

\item{style}{a character vector indicating the font style, i.e.,
\code{"regular"} (default) for regular text, \code{"bold"}
for bold text, and \code{"italic"} for italic text. Note
that the font style \code{"bold"} and \code{"italic"} can
be combined, i.e., style = c("bold", "italic") provides a
bold and italic text. Note that the argument \code{color}
needs to be specified to change the style of the text, e.g.
\code{color = "black"} and \code{style = "bold"} to for
bold text.}

\item{na.omit}{logical: if \code{TRUE}, incomplete cases are removed
before conducting the analysis (i.e., listwise deletion).}

\item{digits}{an integer value indicating the number of decimal places
to be used to display descriptive statistics, e.g., mean,
median, and standard deviation.}

\item{r.digits}{an integer value indicating the number of decimal places
to be used to display the item difficulty, item-total
correlation, and the coefficient alpha if the item is deleted.}

\item{as.na}{a numeric vector indicating user-defined missing values,
i.e. these values are converted to \code{NA} before conducting
the analysis.}

\item{write}{a character string naming a file for writing the output into
either a text file with file extension \code{".txt"} (e.g.,
\code{"Output.txt"}) or Excel file with file extension
\code{".xlsx"}  (e.g., \code{"Output.xlsx"}). If the file
name does not contain any file extension, an Excel file will
be written.}

\item{append}{logical: if \code{TRUE} (default), output will be appended
to an existing text file with extension \code{.txt} specified
in \code{write}, if \code{FALSE} existing text file will be
 overwritten.}

\item{check}{logical: if \code{TRUE} (default), argument specification
is checked.}

\item{output}{logical: if \code{TRUE} (default), output is shown.}
}
\value{
Returns an object of class \code{misty.object}, which is a list with following
entries:

\item{\code{call}}{function call}
\item{\code{type}}{type of analysis}
\item{\code{data}}{data frame including all variables used in the analysis}
\item{\code{key}}{list with the answer key and the position of the answer key
                  for each item}
\item{\code{args}}{specification of function arguments}
\item{\code{result}}{result table}
}
\description{
This function conducts distractor analysis for multiple-choice test items,
i.e., computes attractor-distractor selection frequency, percentage, and total
point-biserial or biserial correlation.
}
\note{
This function uses a modified copy of the \code{polyserial()} function in the
\pkg{polycor} package by John Fox when requesting distractor-total biserial
correlation.
}
\examples{
\dontrun{

#————————————————————————————————————————————————————————————————————————————
# Multiple-Choice Items

# Answer key
key <- c("D", "B", "C", "C", "B", "B", "B", "A", "A", "D")

# Example 1a: Corrected distractor-total point-biserial correlation
item.distract(data.items, +mitem, key = key)

# Example 1b: Corrected distractor-total biserial correlation
item.distract(data.items, +mitem, key = key, method = "biser")

# Example 1c: Keep missing values, i.e., do not convert NA into 0
item.distract(data.items, +mitem, key = key, na.keep = TRUE)

# Example 1d: Print corrected distractor-total point-biserial correlation only
item.distract(data.items, +mitem, key = key, print = "r")

# Example 1e: Highlight attractor frequency, percentage, and correlation in bright red
item.distract(data.items, +mitem, key = key, na.keep = TRUE, color = "b.red")

# Example 1f: Highlight attractor frequency, percentage, and correlation in boldface
item.distract(data.items, +mitem, key = key, na.keep = TRUE, color = "black",
              style = "bold")

#————————————————————————————————————————————————————————————————————————————
# Write Results

# Example 2a: Write Results into a text file
item.distract(data.items, +mitem, key = key, write = "Item-Distract.txt", output = FALSE)

# Example 2b: Write Results into an Excel file
item.distract(data.items, +mitem, key = key, write = "Item-Distracts.xlsx", output = FALSE)
}
}
\references{
Fox, J. (2025). \emph{polycor: Polychoric and polyserial correlations}.
R package version 0.8-2. https://doi.org/10.32614/CRAN.package.polycor
}
\seealso{
\code{\link{item.stats}}, \code{\link{item.alpha}}, \code{\link{item.omega}},
\code{\link{item.cfa}}, \code{\link{item.reverse}}, \code{\link{item.scores}}
}
\author{
Takuya Yanagida
}
