% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item.stats.R
\name{item.stats}
\alias{item.stats}
\title{Item Analysis for Dichotomous, Polytomous, and Continuous Items}
\usage{
item.stats(data, ..., exclude = NULL, correct = TRUE,
           method = c("pearson", "pbiser", "biser", "polyser"),
           adjust = c("none", "joint", "approx"),
           missing = c("listwise", "pairwise", "fiml"),
           alternative = c("two.sided", "less", "greater"),
           conf.level = 0.95, na.omit = FALSE, digits = 2, r.digits = 3,
           as.na = NULL, write = NULL, append = TRUE,
           check = TRUE, output = TRUE)
}
\arguments{
\item{data}{a data frame.  Note that dichotomous items must be coded
with \code{0} and \code{1}, while polytomous should be
coded either as ordered factors or as integer values.
However, the function assumes continuous data whenever items
have more than two distinct values. Polytomous items can
be specified either by providing a data frame with ordered
factors or by specifying a data frame with numeric vectors
with integer values, along with setting the argument
\code{method} to \code{polyser}.}

\item{...}{an expression indicating the variable names in \code{data}
e.g., \code{item.stat(dat, x1, x2, x3)}. Note that the
operators \code{+}, \code{-}, \code{~}, \code{:}, \code{::},
and \code{!} can also be used to select variables, see
'Details' in the \code{\link{df.subset}} function.}

\item{exclude}{a character vector indicating items to be excluded from
the analysis.}

\item{correct}{logical: if \code{TRUE} (default), the corrected item-total
correlation is computed.}

\item{method}{a character string indicating which correlation coefficient
is used for the item-total correlation, i.e., \code{"pearson"}
for the Pearson product-moment correlation coefficient when
data are continuous, \code{"pbiser"} for the point-biserial
correlation coefficient when data are dichotomous, \code{"biser"}
for the biserial correlation coefficient when data are dichotomous,
and \code{"polyser"} for the polyserial correlation coefficient
when data are polytomous. By default, the corrected item-total
point-biserial correlation (\code{"pbiser"}) is computed for
dichotomous items, corrected item-total polyserial correlation
(\code{"polyser"}) is computed for polytomous items, and
the corrected item-total product-moment correlation
(\code{"pearson"}) is computed for continuous items.}

\item{adjust}{a character string specifying the non-normality adjustment
method (see 'Details' in the \code{\link{ci.cor}} function),
i.e., \code{"none"} for the Fisher \eqn{z'} confidence interval
for the Pearson product-moment correlation coefficient without
non-normality adjustment, \code{"joint"} for the confidence
interval with non-normality adjustment via sample joint moments,
and \code{"approx"} (default) for the confidence interval
with non-normality adjustment via approximate distribution
by skewness and kurtosis. Note that this argument only applies
to the Pearson product-moment correlation coefficient, i.e.,
\code{method = "pearson"}.}

\item{missing}{a character string indicating how to deal with missing data
when computing coefficient alphas if the item is deleted,
i.e., \code{"listwise"} for listwise deletion, \code{"pairwise"}
(default) for pairwise deletion, \code{"fiml"} for full
information maximum likelihood method. Note that the
argument \code{na.omit} switches to \code{TRUE} when
specifying \code{missing = "listwise"}.}

\item{alternative}{a character string specifying the alternative hypothesis
for the confidence intervals for the item-total correlation,
i.e., \code{"two.sided"} (default), \code{"greater"} or
\code{"less"}.}

\item{conf.level}{a numeric value between 0 and 1 indicating the confidence
level of the interval for the item-total correlation.}

\item{na.omit}{logical: if \code{TRUE}, incomplete cases are removed
before conducting the analysis (i.e., listwise deletion).
By default pairwise deletion is used for computing point
biserial and product-moment correlation, while listwise
deletion is used for computing biserial and polyserial
correlation. Note that the argument \code{missing}
switches to \code{"listwise"} when specifying
\code{na.omit = TRUE}.}

\item{digits}{an integer value indicating the number of decimal places
to be used to display display item difficulty, mean, median, and standard deviation.}

\item{r.digits}{an integer value indicating the number of decimal places
to be used to display item-total correlation, and coefficient alpha if
the item is deleted.}

\item{as.na}{a numeric vector indicating user-defined missing values,
i.e. these values are converted to \code{NA} before conducting
the analysis.}

\item{write}{a character string naming a file for writing the output into
either a text file with file extension \code{".txt"} (e.g.,
\code{"Output.txt"}) or Excel file with file extension
\code{".xlsx"}  (e.g., \code{"Output.xlsx"}). If the file
name does not contain any file extension, an Excel file will
be written.}

\item{append}{logical: if \code{TRUE} (default), output will be appended
to an existing text file with extension \code{.txt} specified
in \code{write}, if \code{FALSE} existing text file will be
 overwritten.}

\item{check}{logical: if \code{TRUE} (default), argument specification
is checked.}

\item{output}{logical: if \code{TRUE} (default), output is shown.}
}
\value{
Returns an object of class \code{misty.object}, which is a list with following
entries:

\item{\code{call}}{function call}
\item{\code{type}}{type of analysis}
\item{\code{data}}{data frame including all variables used in the analysis}
\item{\code{dtype}}{type of the data, i.e., \code{"dich"} for dichotomous,
                    \code{"poly"} for polytomous, and \code{"cont"} for
                    continuous data}
\item{\code{args}}{specification of function arguments}
\item{\code{result}}{result table}
}
\description{
This function conducts item analysis for dichotomous, polytomous, and continuous
items by computing corrected item-total correlations with confidence intervals
and coefficient alphas if the item is deleted. The default setting for computing
the item-total correlations is depending on the type of the data, i.e.,
point-biserial correlation for dichotomous data, polyserial correlation for
polytomous data, and product-moment correlation for continuous data.
}
\details{
\describe{
\item{\strong{Confidence Interval for the Item-Total Correlation}}{The confidence
interval for the point-biserial correlation is based on Bonett (2020), while
the confidence interval for the biserial and polyserial correlation is based
on the maximum likelihood standard error provided by the function \code{polyserial}
from the \pkg{polycor} package (Fox, 2025). The confidence interval for the
Pearson product-moment correlation uses the \code{ci.cor} function from the
\pkg{misty} package that computes a confidence interval with non-normality
adjustment via approximate distribution by skewness and kurtosis by default.}
}
}
\note{
This function uses a modified copy of the \code{polyserial()} function in the
\pkg{polycor} package by John Fox (2025) when requesting item-total biserial
correlation.
}
\examples{
\dontrun{

#————————————————————————————————————————————————————————————————————————————
# Dichotomous Data

# Example 1a: Corrected item-total point-biserial correlation
item.stats(data.items, +ditem)

# Example 1b: Corrected item-total biserial correlation
item.stats(data.items, +ditem, method = "biser")

# Example 1c: Uncorrected item-total point-biserial correlation
item.stats(data.items, +ditem, correct = FALSE)

# Example 1d: Display item-total correlation and coefficient alpha with 2 digits
item.stats(data.items, +ditem, r.digits = 2)

#————————————————————————————————————————————————————————————————————————————
# Polytomous Data

# Reverse code inverted item
data.items <- item.reverse(data.items, pitem2, pitem3, min = 0, max = 3)

# Example 2a: Corrected item-total polyserial correlation
item.stats(data.items, pitem1, pitem2.r, pitem3.r, pitem4::pitem6, method = "polyser")

# Example 2b: One-sided confidence intervals
item.stats(data.items, +pitem, method = "polyser", alternative = "greater")

#————————————————————————————————————————————————————————————————————————————
# Continuous Data

# Load data set "HolzingerSwineford1939" in the lavaan package
data("HolzingerSwineford1939", package = "lavaan")

# Example 3a: Corrected item-total product-moment correlation
item.stats(HolzingerSwineford1939, x1::x9)

# Example 3a: Exclude items 'x2' and 'x7'
item.stats(HolzingerSwineford1939, x1::x9, exclude = c("x2", "x7"))

#————————————————————————————————————————————————————————————————————————————
# Write Results

# Example 4a: Write Results into a text file
item.stats(data.items, +ditem, write = "Item-Stats.txt", output = FALSE)

# Example 4b: Write Results into an Excel file
item.stats(data.items, +ditem, write = "Item-Stats.xlsx", output = FALSE)
}
}
\references{
Bonett D. G. (2020). Point-biserial correlation: Interval estimation, hypothesis
testing, meta-analysis, and sample size determination. \emph{The British Journal
of Mathematical and Statistical Psychology, 73} Suppl 1, 113–144.
https://doi.org/10.1111/bmsp.12189

Fox, J. (2025). \emph{polycor: Polychoric and polyserial correlations}.
R package version 0.8-2. https://doi.org/10.32614/CRAN.package.polycor
}
\seealso{
\code{\link{item.distract}}, \code{\link{item.alpha}}, \code{\link{item.omega}},
\code{\link{item.cfa}}, \code{\link{item.reverse}}, \code{\link{item.scores}}
}
\author{
Takuya Yanagida
}
