% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multilevel.alpha.R
\name{multilevel.alpha}
\alias{multilevel.alpha}
\title{Multilevel Coefficient Alpha}
\usage{
multilevel.alpha(data, ..., cluster,
                 se = c("none", "standard", "robust.huber.white"),
                 optim.method = c("nlminb", "em"), missing = c("listwise", "fiml"),
                 nrep = 100000, seed = NULL, conf.level = 0.95,
                 print = c("all", "alpha", "item"), digits = 2, r.digits = 3,
                 as.na = NULL, write = NULL, append = TRUE, check = TRUE,
                 output = TRUE)
}
\arguments{
\item{data}{a data frame. A saturated multilevel model is estimated to
obtain variances and covariances at the Within and Between
level. Note that the cluster variable specified in \code{cluster}
is excluded from \code{data} when specifying the argument
\code{cluster} using the variable name of the cluster variable.}

\item{...}{an expression indicating the variable names in \code{data},
e.g., \code{multilevel.alpha(dat, x1, x2, x3, cluster = "cluster")}.
Note that the operators \code{+}, \code{-}, \code{~}, \code{:},
\code{::}, and \code{!} can also be used to select variables,
see 'Details' in the \code{\link{df.subset}} function.}

\item{cluster}{either a character string indicating the variable name of
the cluster variable in \code{data}, or a vector representing
the nested grouping structure (i.e., group or cluster variable).}

\item{se}{a character string indicating the standard errors used for
computing Monte Carlo confidence intervals, i.e., \code{"none"}
for no standard errors, \code{"standard"} for conventional
standard error based on inverting the expected observed or
first.order information matrix, and \code{"robust.huber.white"}
for the 'MLR' (aka pseudo ML, Huber-White) approach. Note
that \code{se = "none"} saves computation time by not
computing standard errors and Monte Carlo confidence intervals.}

\item{optim.method}{a character string indicating the optimizer, i.e., \code{"nlminb"}
(default) for the unconstrained and bounds-constrained
quasi-Newton method optimizer and \code{"em"} for the
Expectation Maximization (EM) algorithm.}

\item{missing}{a character string indicating how to deal with missing data,
i.e., \code{"listwise"} for listwise deletion or \code{"fiml"}
(default) for full information maximum likelihood (FIML)
method.}

\item{nrep}{an integer value indicating the number of Monte Carlo
repetitions for computing confidence intervals.}

\item{seed}{a numeric value specifying the seed of the random number
generator for computing the Monte Carlo confidence interval.}

\item{conf.level}{a numeric value between 0 and 1 indicating the confidence
level of the interval.}

\item{print}{a character vector indicating which results to show, i.e.
\code{"all"} (default) for all results \code{"alpha"} for
the composite reliability alpha and \code{"item"} for item
statistics. Note that standardized factor loadings at the
within and between level printed when requesting \code{"item"}
are based on a multilevel CFA model with factor loadings
freely estimated across both levels.}

\item{digits}{an integer value indicating the number of decimal places
to be used for displaying mean, standard deviation, minimum,
maximum, skewness, and kurtosis.}

\item{r.digits}{an integer value indicating the number of decimal places
to be used for displaying multilevel coefficient alpha,
ICC(1), and standardized factor loadings.}

\item{as.na}{a numeric vector indicating user-defined missing values,
i.e. these values are converted to \code{NA} before conducting
the analysis. Note that \code{as.na()} function is only
applied to \code{data} but not to \code{cluster}.}

\item{write}{a character string naming a file for writing the output into
either a text file with file extension \code{".txt"} (e.g.,
\code{"Output.txt"}) or Excel file with file extension
\code{".xlsx"}  (e.g., \code{"Output.xlsx"}). If the file
name does not contain any file extension, an Excel file will
be written.}

\item{append}{logical: if \code{TRUE} (default), output will be appended
to an existing text file with extension \code{.txt} specified
in \code{write}, if \code{FALSE} existing text file will be
overwritten.}

\item{check}{logical: if \code{TRUE} (default), argument specification,
convergence and model identification is checked.}

\item{output}{logical: if \code{TRUE} (default), output is shown.}
}
\value{
Returns an object of class \code{misty.object}, which is a list with following
entries:
\item{\code{call}}{function call}
\item{\code{type}}{type of analysis}
\item{\code{data}}{data frame specified in \code{data} including the group variable
                   specified in \code{cluster}}
\item{\code{args}}{specification of function arguments}
\item{\code{model}}{specified model}
\item{\code{model.fit}}{fitted lavaan object (\code{mod.fit})}
\item{\code{check}}{results of the convergence and model identification check}
\item{\code{result}}{list with result tables, i.e., \code{alpha} for the coefficient
                     alpha including Monte Carlo confidence interval and
                     \code{itemstat} for descriptive statistics}
}
\description{
This function computes point estimate and Monte Carlo confidence interval for
the multilevel coefficient alpha defined by Lai (2021) by calling the \code{cfa}
function in the R package \pkg{lavaan}. By default, the function prints level-specific
multilevel coefficient alpha with 95\% Monte Carlo confidence interval based on
Huber-White standard errors.
}
\details{
In single-level data, coefficient \eqn{\alpha} (Cronbach, 1951) can be computed
as

\deqn{\alpha = \frac{2p\Sigma_{k=2}^p\Sigma_{k^{\prime}=1}^{k-1}\sigma_{kk^{\prime}}}{(p - 1)\mathbf{1}^{\prime}\mathbf{\Sigma}\mathbf{1}}}

assuming that the covariance matrix of \eqn{p} items is \eqn{\Sigma} with
elements \eqn{\sigma_{ij^{\prime}}}. Note that coefficients \eqn{\alpha} and \eqn{\omega}
represent the same population quantities under unidimensionality and essential
tau-equivalence. The multilevel extension of coefficient \eqn{\alpha}, however,
does not require estimation of parameters of a factor model. Conseuqntly, its
computation at the between-level remains the same regardless a composite is used
to measure a configural construct or a shared construct.

\describe{

  \item{Within-Level Composite}{The reliability for an observed within-level
  composite, \eqn{Z^w_{ij}}, is

  \deqn{\alpha^w = \frac{2p\Sigma_{k=2}^p\Sigma_{k^{\prime}=1}^{k-1}\sigma_{kk^{\prime}}^w}{(p - 1)\mathbf{1}^{\prime}\mathbf{\Sigma}^w\mathbf{1}}}

  }

  \item{Between-Level Composite}{The reliability for an observed between-level
  composite, \eqn{Z^b_{j}}, is

  \deqn{\alpha^b = \frac{2p\Sigma_{k=2}^p\Sigma_{k^{\prime}=1}^{k-1}\sigma_{kk^{\prime}}^b}{(p - 1)[\mathbf{1}^{\prime}\mathbf{\Sigma}^b\mathbf{1} + \mathbf{1}^{\prime}\mathbf{\Sigma}^w\mathbf{1} / \tilde{n}]}}

  }

  \item{Overall Composite}{The reliability for an overall observed composite,
  \eqn{Z_{ij}}, capturing the population variance of both the within-level and the
  between-level components of the true score and the errors is

  \deqn{\alpha^{2l} = \frac{2p\Sigma_{k=2}^p\Sigma_{k^{\prime}=1}^{k-1}(\sigma_{kk^{\prime}}^w + \sigma_{kk^{\prime}}^b) }{(p - 1)[\mathbf{1}^{\prime}\mathbf{\Sigma}^b\mathbf{1} + \mathbf{1}^{\prime}\mathbf{\Sigma}^w\mathbf{1}]}}

  Note that \eqn{\alpha^{2l}} is simply the ratio of true variance to total variance
  of the observed scores. In practice, \eqn{\alpha^{2l}} is not the most interesting
  coefficient because researchers are mainly interested in distinguishing within-
  and between effects (Castro-Alvarez et al., 2026).}}
}
\note{
This function is based on the function \code{multilevel_alpha} from Mark Lai
(2021 supplementary codes) and  uses the functions \code{lavInspect}, \code{lavTech},
\code{lavNames}, \code{parameterEstimates}, and \code{sem} provided in the R
package \pkg{lavaan} by Yves Rosseel (2012). The internal function \code{.internal.mvrnorm}
is a copy of the \code{mvrnorm} function in the package \pkg{MASS} by Venables
and Ripley (2002).
}
\examples{
\dontrun{
# Load data set "Demo.twolevel" in the lavaan package
data("Demo.twolevel", package = "lavaan")

#————————————————————————————————————————————————————————————————————————————
# Cluster Variable Specification

# Example 1a: Specification using the argument '...'
multilevel.alpha(Demo.twolevel, y1:y4, cluster = "cluster")

# Example 1b: Alternative specification with cluster variable 'cluster' in 'data'
multilevel.alpha(Demo.twolevel[, c("y1", "y2", "y3", "y4", "cluster")], cluster = "cluster")

# Example 1c: Alternative specification with cluster variable 'cluster' not in 'data'
multilevel.alpha(Demo.twolevel[, c("y1", "y2", "y3", "y4")], cluster = Demo.twolevel$cluster)

#————————————————————————————————————————————————————————————————————————————
# Arguments 'se' and 'print'

# Example 4a: No confidence intervals to speed up computation
multilevel.alpha(Demo.twolevel, y1:y4, cluster = "cluster", se = "none")

# Example 4b: Alpha and item statistics
multilevel.alpha(Demo.twolevel, y1:y4, cluster = "cluster", print = "all")

#————————————————————————————————————————————————————————————————————————————
# Write Results

# Example 5a: Write results into a text file
multilevel.alpha(Demo.twolevel, y1, y2, y3, y4, cluster = "cluster",
                 write = "Multilevel_Alpha.txt")

# Example 5b: Write results into a Excel file
multilevel.alpha(Demo.twolevel, y1, y2, y3, y4, cluster = "cluster",
                 write = "Multilevel_Alpha.xlsx")
}
}
\references{
Cronbach, L. J. (1951). Coefficient alpha and the internal structure of tests.
\emph{Psychometrika, 16}, 297-334. http://dx.doi.org/10.1007/BF02310555

Lai, M. H. C. (2021). Composite reliability of multilevel data: It’s about
observed scores and construct meanings. \emph{Psychological Methods, 26}(1),
90–102. https://doi.org/10.1037/met0000287

Rosseel, Y. (2012). lavaan: An R Package for Structural Equation Modeling.
\emph{Journal of Statistical Software, 48}, 1-36. https://doi.org/10.18637/jss.v048.i02

Venables, W. N., Ripley, B. D. (2002).\emph{Modern Applied Statistics with S} (4th ed.).
Springer. https://www.stats.ox.ac.uk/pub/MASS4/.
}
\seealso{
\code{\link{multilevel.omega}}, \code{\link{item.omega}}, \code{\link{multilevel.cfa}},
\code{\link{multilevel.fit}}, \code{\link{multilevel.invar}}, \code{\link{multilevel.cor}},
\code{\link{multilevel.descript}}, \code{\link{write.result}}
}
\author{
Takuya Yanagida \email{takuya.yanagida@univie.ac.at}
}
