% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.misty.object.R
\name{print.misty.object}
\alias{print.misty.object}
\title{Print misty.object object}
\usage{
\method{print}{misty.object}(x, sort.var = x$args$sort.var,
      digits = x$args$digits, p.digits = x$args$p.digits,
      icc.digits = x$args$icc.digits, r.digits = x$args$r.digits,
      ess.digits = x$args$ess.digits, mcse.digits = x$args$mcse.digits,
      check = TRUE, ...)
}
\arguments{
\item{x}{\code{misty.object} object.}

\item{sort.var}{logical: if \code{TRUE}, output is sorted by variables.}

\item{digits}{an integer value indicating the number of decimal places
to be used for displaying results.}

\item{p.digits}{an integer indicating the number of decimal places to be used
for displaying \emph{p}-values.}

\item{icc.digits}{an integer indicating the number of decimal places to be used
for displaying intraclass correlation coefficients
(\code{multilevel.descript()} and \code{multilevel.icc()}
function).}

\item{r.digits}{an integer value indicating the number of decimal places
to be used for displaying R-hat values, item-total correlation
standardized factor loadings, coefficient alpha, and
coefficient omega.}

\item{ess.digits}{an integer value indicating the number of decimal places
to be used for displaying effective sample sizes.}

\item{mcse.digits}{an integer value indicating the number of decimal places
to be used for displaying Monte Carlo standard errors.}

\item{check}{logical: if \code{TRUE}, argument specification is checked.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
This function prints an \code{misty.object} object.
}
\examples{
\dontrun{
#————————————————————————————————————————————————————————————————————————————
# Example 1: item.cfa() function

# Load data set "HolzingerSwineford1939" in the lavaan package
data("HolzingerSwineford1939", package = "lavaan")

result <- item.cfa(HolzingerSwineford1939[, c("x1", "x2", "x3")], output = FALSE)
print(result)

#————————————————————————————————————————————————————————————————————————————
# Example 2: multilevel.descript() function

# Load data set "Demo.twolevel" in the lavaan package
data("Demo.twolevel", package = "lavaan")

result <- multilevel.descript(y1:y3, data = Demo.twolevel, cluster = "cluster",
                              output = FALSE)
print(result)
}
}
\author{
Takuya Yanagida \email{takuya.yanagida@univie.ac.at}
}
