% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mplus.plot.R
\name{mplus.plot}
\alias{mplus.plot}
\title{Plot Mplus GH5 File}
\usage{
mplus.plot(x, plot = c("none", "trace", "post", "auto", "ppc", "loop"),
           param = c("all", "on", "by", "with", "inter", "var", "r2", "new"),
           std = c("all", "none", "stdyx", "stdy", "std"), burnin = TRUE,
           point = c("all", "none", "m", "med", "map"),
           ci = c("none", "eti", "hdi"), chain = 1, conf.level = 0.95,
           hist = TRUE, density = TRUE, area = TRUE, alpha = 0.4,
           fill = "gray85", facet.nrow = NULL, facet.ncol = NULL,
           facet.scales = c("fixed", "free", "free_x", "free_y"),
           xlab = NULL, ylab = NULL, xlim = NULL, ylim = NULL,
           xbreaks = ggplot2::waiver(), ybreaks = ggplot2::waiver(),
           xexpand = ggplot2::waiver(), yexpand = ggplot2::waiver(),
           palette = "Set 2", binwidth = NULL, bins = NULL,
           density.col = "#0072B2", shape = 21,
           point.col = c("#CC79A7", "#D55E00", "#009E73"),
           linewidth = 0.6, linetype = "dashed", line.col = "black",
           bar.col = "black", bar.width = 0.8, plot.margin = NULL,
           legend.title.size = 10, legend.text.size = 10, legend.box.margin = NULL,
           saveplot = c("all", "none", "trace", "post", "auto", "ppc", "loop"),
           filename = "Mplus_Plot.pdf",
           file.plot = c("_TRACE", "_POST", "_AUTO", "_PPC", "_LOOP"),
           width = NA, height = NA, units = c("in", "cm", "mm", "px"),
           dpi = 600, check = TRUE)
}
\arguments{
\item{x}{a character string indicating the name of the Mplus
GH5 file (HDF5 format) with or without the file
extension \code{.gh5}, e.g., \code{"Mplus_Plot.gh5"}
or \code{"Mplus_Plot"}. Alternatively, a \code{misty.object}
of type \code{mplus} can be specified, i.e., result
object of the \code{mplus.plot()} function.}

\item{plot}{a character string indicating the type of plot to
display, i.e., \code{"none"} for not displaying any
plot, \code{"trace"} (default) for displaying trace
plots, \code{post} for displaying posterior distribution
plots, \code{"auto"} for displaying autocorrelation
plots, \code{"ppc"} for displaying posterior predictive
check plots, and \code{"loop"} for displaying the
loop plot. The default setting is \code{"trace"} if
the "bayesian_data" section is available in the Mplus
GH5 file. Otherwise, the default setting switches to
\code{"loop"}.}

\item{param}{character vector indicating which parameters to print
for the trace plots, posterior distribution plots,
and autocorrelation plots, i.e., \code{"all"} for all
parameters, \code{"on"} (default), for regression
slopes, \code{"by"} for factor loadings, \code{"with"}
for covariances, \code{"inter"} for intercepts and
thresholds, \code{"var"} for (residual) variances,
\code{"r2"} for r-square, and \code{"new"} for
parameters not in the analysis model specified in the
\code{NEW} option. The default setting is \code{"on"}
if regression slopes are available. Otherwise, the
default setting switches to \code{"by"} and to
\code{"with"} if factor loadings are not available.}

\item{std}{a character vector indicating the standardized
parameters to print for the trace plots, posterior
distribution plots, and autocorrelation plots, i.e.,
\code{"all"} for all standardized parameters,
\code{"none"} (default) for not printing any
standardized parameters, \code{"stdyx"} for StdYX
standardized parameters, \code{"stdy"} for StdY
standardized parameters, and \code{"std"} for StdX
standardized parameters.}

\item{burnin}{logical: if \code{FALSE}, the first half of each chain
is discarded as being part of the burnin phase when
displaying trace plots. The default setting for
\code{plot = "trace"} is \code{TRUE}. Note that the
first half of each chain is always discarded when
displaying posterior distribution plots (\code{plot = "post"}) regardless
of the setting of the argument \code{burnin}.}

\item{point}{a character vector indicating the point estimate(s)
to be displayed in the posterior distribution plots,
i.e., \code{"all"} for all point estimates, \code{"none"}
for not displaying any point estimates, \code{"m"}
for the posterior mean estimate, \code{"med"} (default)
for the posterior median estimate, and \code{"map"}
for the maximum a posteriori estimate.}

\item{ci}{a character string indicating the type of credible
interval to be displayed in the posterior distribution
plots, i.e., \code{"none"} for not displaying any
credible intervals, \code{"eti"} (default) for displaying
the equal-tailed intervals and \code{"hdi"} for displaying
the highest density interval.}

\item{chain}{a numerical value indicating the chain to be used for
the autocorrelation plots. By default, the first
chain is used.}

\item{conf.level}{a numeric value between 0 and 1 indicating the
confidence level of the credible interval (default is
\code{0.95}).}

\item{hist}{logical: if \code{TRUE} (default), histograms are
drawn in the posterior probability plots.}

\item{density}{logical: if \code{TRUE} (default), density curves are
drawn in the posterior probability plots.}

\item{area}{logical: if \code{TRUE} (default), statistical not
significant and statistical significant area is
filled with a different color and vertical lines are
drawn.}

\item{alpha}{a numeric value between 0 and 1 for the \code{alpha}
argument (default is \code{0.4}) for the \code{annotate},
\code{geom_histogram}, \code{geom_bar}, and
\code{geom_ribbon} function.}

\item{fill}{a character string indicating the color for the
\code{"fill"} argument (default is \code{"gray85"})
for the \code{annotate}, \code{geom_histogram},
\code{geom_bar}, and \code{geom_point} functions.}

\item{facet.nrow}{a numeric value indicating the \code{nrow} argument
(default is \code{NULL}) for the \code{facet_wrap}
function.}

\item{facet.ncol}{a numeric value indicating the \code{ncol} argument
(default is \code{2}) for the \code{facet_wrap} function.}

\item{facet.scales}{a character string indicating the \code{scales} argument
(default is \code{"free"}) for the \code{facet_wrap}
function.}

\item{xlab}{a character string indicating the \code{name} argument
for the \code{scale_x_continuous} function. Note that
the default setting depends on the type of plot,
e.g., \code{""} for the trace plots and \code{"Lag"}
for the autocorrelation plots.}

\item{ylab}{a character string indicating the \code{name} argument
for the \code{scale_y_continuous} function. Note that
the default setting depends on the type of plot,
e.g., \code{""} for the trace plots and \code{"Autocorrelation"}
for the autocorrelation plots.}

\item{xlim}{a numeric vector with two elements indicating the
\code{limits} argument (default it \code{NULL}) for
the \code{scale_x_continuous} function.}

\item{ylim}{a numeric vector with two elements indicating the
\code{limits} argument (default it \code{NULL}) for
the \code{scale_y_continuous} function.}

\item{xbreaks}{a numeric vector indicating the \code{breaks} argument
(default is \code{ggplot2::waiver()}) for the
\code{scale_x_continuous} function.}

\item{ybreaks}{a numeric vector indicating the \code{breaks} argument
(default is \code{ggplot2::waiver()}) for the
\code{scale_y_continuous} function.}

\item{xexpand}{a numeric vector with two elements indicating the
\code{expand} argument (default is \code{(0.02, 0)})
for the \code{scale_x_continuous} function.}

\item{yexpand}{a numeric vector with two elements indicating the
\code{expand} argument for the \code{scale_y_continuous}
function. Note that the default setting depends
on the type of plot, e.g., \code{(0.02, 0)} for the
trace plots and \code{expansion(mult = c(0, 0.05))}
for the posterior distribution plots.}

\item{palette}{a character string indicating the palette name (default
is \code{"Set 2"}) for the \code{hcl.colors} function.
Note that the character string must be one of
\code{hcl.pals()}.}

\item{binwidth}{a numeric value indicating the \code{binwidth} argument
(default is to use the number of bins in \code{bins}
argument) for the \code{geom_histogram} function.}

\item{bins}{a numeric value indicating the \code{bins} argument
(default is \code{30}) for the \code{geom_histogram}
function.}

\item{density.col}{a character string indicating the \code{color} argument
(default is \code{"#0072B2"}) for the \code{geom_density}
function.}

\item{shape}{a numeric value indicating the \code{shape} argument
(default is \code{21}) for the \code{geom_point}
function.}

\item{point.col}{a character vector with three elements indicating the
\code{values} argument (default is
\code{c("#CC79A7", "#D55E00", "#009E73")}) for the
\code{scale_color_manual} function.}

\item{linewidth}{a numeric value indicating the \code{linewidth} argument
(default is \code{0.6}) for the \code{geom_vline} function.}

\item{linetype}{a numeric value indicating the \code{linetype} argument
(default is \code{"dashed"}) for the \code{geom_vline}
function.}

\item{line.col}{a character string indicating the \code{color} argument
(default is \code{"black"}) for the \code{geom_vline}
function.}

\item{bar.col}{a character string indicating the \code{color} argument
(default is \code{"black"}) for the \code{geom_bar}
function.}

\item{bar.width}{a character string indicating the \code{width} argument
(default = \code{0.8})for the \code{geom_bar} function.}

\item{plot.margin}{a numeric vector indicating the \code{plot.margin}
argument for the \code{theme} function. Note that the
default setting depends on the type of the plot, e.g.,
\code{c(4, 15, -10, 0)} for the trace plots, and
\code{c(4, 15, 4, 4)} for the autocorrelation plots.}

\item{legend.title.size}{a numeric value indicating the \code{legend.title}
argument (default is \code{element_text(size = 10)})
for the \code{theme} function.}

\item{legend.text.size}{a numeric value indicating the \code{legend.text}
argument (default is \code{element_text(size = 10)})
for the \code{theme} function.}

\item{legend.box.margin}{a numeric vector indicating the \code{legend.box.margin}
argument for the \code{theme} function. Note that the
default setting depends on the type of plot, e.g.,
\code{c(-16, 6, 6, 6)} for the trace plots, and
\code{c(-25, 6, 6, 6)} for the posterior distribution
plots with displaying point estimates.}

\item{saveplot}{a character vector indicating the plot to be saved,
i.e., \code{"all"} for saving all plots, \code{"none"}
(default) for not saving any plots, \code{"trace"}
for saving the trace plots, \code{post} for the saving
the posterior distribution plots, \code{"auto"} for
saving the autocorrelation plots, \code{"ppc"} for
saving the posterior predictive check plots, and
\code{"loop"} for saving the loop plot.}

\item{filename}{a character string indicating the \code{filename}
argument (default is \code{"Mplus_Plot.pdf"}) including
the file extension for the \code{ggsave} function.
Note that one of \code{".eps"}, \code{".ps"},
\code{".tex"}, \code{".pdf"} (default), \code{".jpeg"},
\code{".tiff"}, \code{".png"}, \code{".bmp"},
\code{".svg"} or \code{".wmf"} needs to be specified
as file extension in the \code{filename} argument.}

\item{file.plot}{a character vector with five elements for distinguishing
different types of plots. By default, the character
string specified in the argument \code{"filename"}
(\code{"Mplus_Plot"}) is concatenated with \code{"_TRACE"}
(\code{"Mplus_Plot_TRACE"}) for the trace plots,
\code{"_POST"} (\code{"Mplus_Plot_POST"}) for
the posterior distribution plots, \code{"_AUTO"}
(\code{"Mplus_Plot_AUTO"}) for the autocorrelation
plots, \code{"_PPC"} (\code{"Mplus_Plot_PPC"}) for the
posterior predictive check plots, and \code{"_LOOP"}
(\code{"Mplus_Plot_LOOP"}) for the loop plot.}

\item{width}{a numeric value indicating the \code{width} argument
(default is the size of the current graphics device)
for the \code{ggsave} function.}

\item{height}{a numeric value indicating the \code{height} argument
(default is the size of the current graphics device)
for the \code{ggsave} function.}

\item{units}{a character string indicating the \code{units}
argument (default is \code{in}) for the \code{ggsave}
function.}

\item{dpi}{a numeric value indicating the \code{dpi} argument
(default is \code{600}) for the \code{ggsave} function.}

\item{check}{logical: if \code{TRUE} (default), argument
specification is checked.}
}
\value{
Returns an object of class \code{misty.object}, which is a list with following
entries:

\item{\code{call}}{function call}
\item{\code{type}}{type of analysis}
\item{\code{x}}{Mplus GH5 file}
\item{\code{args}}{specification of function arguments}
\item{\code{data}}{list with posterior distribution of each parameter estimate
                   in wide and long format (\code{post}), autocorrelation for
                   each parameter estimate in wide and long format (\code{auto}),
                   data for the posterior predictive check (\code{ppc}),
                   and data for the loop plot (\code{loop})}
\item{\code{plot}}{list with the trace plots (\code{trace}, posterior distribution
                   plots (\code{post}), autocorrelation plots (\code{auto}),
                   posterior predictive check plots (\code{ppc}), and
                   loop plot (\code{loop})}
}
\description{
This function uses the \code{h5file} function in the \pkg{hdf5r} package to
read a Mplus GH5 file that is requested by the command \code{PLOT: TYPE IS PLOT2}
in Mplus to display trace plots, posterior distribution plots, autocorrelation
plots, posterior predictive check plots based on the "bayesian_data" section, and
the loop plot based on the "loop_data" section of the Mplus GH5 file. By default,
the function displays trace plots if the "bayesian_data" section is available in
the Mplus GH5 File. Otherwise, the function plots the loop plot if the "loop_data"
section is available in the Mplus GH5 file.
}
\examples{
\dontrun{

#————————————————————————————————————————————————————————————————————————————
# Mplus Example 3.18: Moderated Mediation with a Plot of the Indirect Effect

#···················
# Trace Plots

# Example 1a: Default setting
mplus.plot("ex3.18.gh5")

# Example 1b: Exclude first half of each chain
mplus.plot("ex3.18.gh5", burnin = FALSE)

# Example 1c: Print all parameters
mplus.plot("ex3.18.gh5", param = "all")

# Example 1d: Print user-specified parameters
mplus.plot("ex3.18.gh5", param = "param")

# Example 1e: Arrange panels in three columns
mplus.plot("ex3.18.gh5", ncol = 3)

# Example 1f: Specify "Pastel 1" palette for the hcl.colors function
mplus.plot("ex3.18.gh5", palette = "Pastel 1")

#···················
# Posterior Distribution Plots

# Example 2a: Default setting, i.e., posterior median and equal-tailed interval
mplus.plot("ex3.18.gh5", plot = "post")

# Example 2b: Display posterior mean and maximum a posteriori
mplus.plot("ex3.18.gh5", plot = "post", point = c("m", "map"))

# Example 2c: Display maximum a posteriori and highest density interval
mplus.plot("ex3.18.gh5", plot = "post", point = "map", ci = "hdi")

# Example 2d: Do not display any point estimates and credible interval
mplus.plot("ex3.18.gh5", plot = "post", point = "none", ci = "none")

# Example 2d: Do not display histograms
mplus.plot("ex3.18.gh5", plot = "post", hist = FALSE)

#···················
# Autocorrelation Plots

# Example 3a: Default setting, i.e., first chain
mplus.plot("ex3.18.gh5", plot = "auto")

# Example 3b: Use second chain
mplus.plot("ex3.18.gh5", plot = "auto", chain = 2)

# Example 3b: Modify limits and breaks of the y-axis
mplus.plot("ex3.18.gh5", plot = "auto",
           ylim = c(-0.05, 0.05), ybreaks = seq(-0.1, 0.1, by = 0.025))

#···················
# Posterior Predictive Check Plots

# Example 4a: Default setting, i.e., 95\% Interval
mplus.plot("ex3.18.gh5", plot = "ppc")

# Example 4b: Default setting, i.e., 99\% Interval
mplus.plot("ex3.18.gh5", plot = "ppc", conf.level = 0.99)

#···················
# Loop Plot

# Example 5a: Default setting
mplus.plot("ex3.18.gh5", plot = "loop")

# Example 5b: Do not fill area and draw vertical lines
mplus.plot("ex3.18.gh5", plot = "loop", area = FALSE)

#···················
# Save Plots

# Example 6a: Save all plots in pdf format
mplus.plot("ex3.18.gh5", saveplot = "all")

# Example 6b: Save all plots in png format with 300 dpi
mplus.plot("ex3.18.gh5", saveplot = "all", filename = "Mplus_Plot.png", dpi = 300)

# Example 6a: Save loop plot, specify width and height of the plot
mplus.plot("ex3.18.gh5", plot = "none", saveplot = "loop",
           width = 7.5, height = 7)

#————————————————————————————————————————————————————————————————————————————
# Plot from misty.object

# Create misty.object
object <- mplus.plot("ex3.18.gh5", plot = "none")

# Trace plot
mplus.plot(object, plot = "trace")

# Posterior distribution plot
mplus.plot(object, plot = "post")

# Autocorrelation plot
mplus.plot(object, plot = "auto")

# Posterior predictive check plot
mplus.plot(object, plot = "ppc")

# Loop plot
mplus.plot(object, plot = "loop")

#————————————————————————————————————————————————————————————————————————————
# Create Plots Manually

# Load ggplot2 package
library(ggplot2)

# Create misty object
object <- mplus.plot("ex3.18.gh5", plot = "none")

#···················
# Example 7: Trace Plots

# Extract data in long format
data.post <- object$data$post$long

# Extract ON parameters
data.trace <- data.post[grep(" ON ", data.post$param), ]

# Plot
ggplot(data.trace, aes(x = iter, y = value, color = chain)) +
  annotate("rect", xmin = 0, xmax = 15000, ymin = -Inf, ymax = Inf,
           alpha = 0.4, fill = "gray85") +
  geom_line() +
  facet_wrap(~ param, ncol = 2, scales = "free") +
  scale_x_continuous(name = "", expand = c(0.02, 0)) +
  scale_y_continuous(name = "", expand = c(0.02, 0)) +
  scale_colour_manual(name = "Chain",
                      values = hcl.colors(n = 2, palette = "Set 2")) +
  theme_bw() +
  guides(color = guide_legend(nrow = 1, byrow = TRUE)) +
  theme(plot.margin = margin(c(4, 15, -10, 0)),
        legend.position = "bottom",
        legend.title = element_text(size = 10),
        legend.text = element_text(size = 10),
        legend.box.margin = margin(c(-16, 6, 6, 6)),
        legend.background = element_rect(fill = "transparent"))

#···················
# Example 8: Posterior Distribution Plots

# Extract data in long format
data.post <- object$data$post$long

# Extract ON parameters
data.post <- data.post[grep(" ON ", data.post$param), ]

# Discard burn-in iterations
data.post <- data.post[data.post$iter > 15000, ]

# Drop factor levels
data.post$param <- droplevels(data.post$param,
                              exclude = c("[Y]", "[M]", "Y", "M", "INDIRECT", "MOD"))

# Plot
ggplot(data.post, aes(x = value)) +
  geom_histogram(aes(y = after_stat(density)), color = "black", alpha = 0.4,
                 fill = "gray85") +
  geom_density(color = "#0072B2") +
  geom_vline(data = data.frame(param = unique(data.post$param),
                               stat = tapply(data.post$value, data.post$param, median)),
             aes(xintercept = stat, color = "Median"), linewidth = 0.6) +
  geom_vline(data = data.frame(param = unique(data.post$param),
                               low = tapply(data.post$value, data.post$param,
                                            function(y) quantile(y, probs = 0.025))),
             aes(xintercept = low), linetype = "dashed", linewidth = 0.6) +
  geom_vline(data = data.frame(param = unique(data.post$param),
                               upp = tapply(data.post$value, data.post$param,
                                            function(y) quantile(y, probs = 0.975))),
             aes(xintercept = upp), linetype = "dashed", linewidth = 0.6) +
  facet_wrap(~ param, ncol = 2, scales = "free") +
  scale_x_continuous(name = "", expand = c(0.02, 0)) +
  scale_y_continuous(name = "Probability Density, f(x)",
                     expand = expansion(mult = c(0L, 0.05))) +
  scale_color_manual(name = "Point Estimate", values = c(Median = "#D55E00")) +
  labs(caption = "95\% Equal-Tailed Interval") +
  theme_bw() +
  theme(plot.margin = margin(c(4, 15, -8, 4)),
        plot.caption = element_text(hjust = 0.5, vjust = 7),
        legend.position = "bottom",
        legend.title = element_text(size = 10),
        legend.text = element_text(size = 10),
        legend.box.margin = margin(c(-30, 6, 6, 6)),
        legend.background = element_rect(fill = "transparent"))

#···················
# Example 9: Autocorrelation Plots

# Extract data in long format
data.auto <- object$data$auto$long

# Select first chain
data.auto <- data.auto[data.auto$chain == 1, ]

# Extract ON parameters
data.auto <- data.auto[grep(" ON ", data.auto$param), ]

# Plot
ggplot(data.auto, aes(x = lag, y = cor)) +
  geom_bar(stat = "identity", alpha = 0.4, color = "black", fill = "gray85",
           width = 0.8) +
  facet_wrap(~ param, ncol = 2) +
  scale_x_continuous(name = "Lag", breaks = seq(1, 30, by = 2), expand = c(0.02, 0)) +
  scale_y_continuous(name = "Autocorrelation", limits = c(-0.1, 0.1),
                     breaks = seq(-0.1, 1., by = 0.05), expand = c(0.02, 0)) +
  theme_bw() +
  theme(plot.margin = margin(c(4, 15, 4, 4)))

#···················
# Example 10: Posterior Predictive Check (PPC) Plots

# Extract data
data.ppc <- object$data$ppc

# Scatter plot
ppc.scatter <- ggplot(data.ppc, aes(x = obs, y = rep)) +
  geom_point(shape = 21, fill = "gray85") +
  geom_abline(slope = 1) +
  scale_x_continuous("Observed", limits = c(0, 45), breaks = seq(0, 45, by = 5),
                     expand = c(0.02, 0)) +
  scale_y_continuous("Recpliated", limits = c(0, 45), breaks = seq(0, 45, by = 5),
                     expand = c(0.02, 0)) +
  theme_bw() +
  theme(plot.margin = margin(c(2, 15, 4, 4)))

# Histogram
ppc.hist <- ggplot(data.ppc, aes(x = diff)) +
  geom_histogram(color = "black", alpha = 0.4, fill = "gray85") +
  geom_vline(xintercept = mean(data.ppc$diff), color = "#CC79A7") +
  geom_vline(xintercept = quantile(data.ppc$diff, probs = 0.025),
             linetype = "dashed", color = "#CC79A7") +
  geom_vline(xintercept = quantile(data.ppc$diff, probs = 0.975),
             linetype = "dashed", color = "#CC79A7") +
  scale_x_continuous("Observed - Replicated", expand = c(0.02, 0)) +
  scale_y_continuous("Count", expand = expansion(mult = c(0L, 0.05))) +
  theme_bw() +
  theme(plot.margin = margin(c(2, 15, 4, 4)))

# Combine plots using the patchwork package
patchwork::wrap_plots(ppc.scatter, ppc.hist)

#···················
# Example 11: Loop Plot

# Extract data
data.loop <- object$data$loop

# Plot
plot.loop <- ggplot(data.loop, aes(x = xval, y = estimate)) +
  geom_line(linewidth = 0.6, show.legend = FALSE) +
  geom_line(aes(xval, low)) +
  geom_line(aes(xval, upp)) +
  scale_x_continuous("MOD", expand = c(0.02, 0)) +
  scale_y_continuous("INDIRECT", expand = c(0.02, 0)) +
  scale_fill_manual("Statistical Significance",
                    values = hcl.colors(n = 2, palette = "Set 2")) +
  theme_bw() +
  theme(plot.margin = margin(c(4, 15, -6, 4)),
        legend.position = "bottom",
        legend.title = element_text(size = 10),
        legend.text = element_text(size = 10),
        legend.box.margin = margin(-10, 6, 6, 6),
        legend.background = element_rect(fill = "transparent"))

# Significance area
for (i in unique(data.loop$group)) {

  plot.loop <- plot.loop + geom_ribbon(data = data.loop[data.loop$group == i, ],
                                       aes(ymin = low, ymax = upp, fill = sig), alpha = 0.4)

}

# Vertical lines
plot.loop + geom_vline(data = data.loop[data.loop$change == 1, ],
                       aes(xintercept = xval, color = sig), linewidth = 0.6,
                           linetype = "dashed", show.legend = FALSE)
}
}
\references{
Muthen, L. K., & Muthen, B. O. (1998-2017). \emph{Mplus User's Guide} (8th ed.).
Muthen & Muthen.
}
\seealso{
\code{\link{read.mplus}}, \code{\link{write.mplus}}, \code{\link{mplus}},
\code{\link{mplus.update}}, \code{\link{mplus.print}}, \code{\link{mplus.bayes}},
\code{\link{mplus.run}}, \code{\link{mplus.lca}}
}
\author{
Takuya Yanagida
}
