% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.misty.object.R
\name{plot.misty.object}
\alias{plot.misty.object}
\title{Plots misty.object object}
\usage{
\method{plot}{misty.object}(x, plot = x$args$plot, bar = x$args$bar,
     box = x$args$box, violin = x$args$violin, hist = x$args$hist,
     point = x$args$point, line = x$args$line, ci = x$args$ci,
     conf.level = x$args$conf.level, adjust = x$args$adjust,
     jitter = x$args$jitter, density = x$args$density,
     square = x$args$square, rotate = x$args$rotate,
     binwidth = x$args$binwidth, bins = x$args$bins,
     fill = x$args$fill, hist.alpha = x$args$hist.alpha,
     tile.alpha = x$args$tile.alpha, violin.alpha = x$args$violin.alpha,
     violin.trim = x$args$violin.trim, box.width = x$args$box.width,
     box.alpha = x$args$box.alpha, linetype = x$args$linetype,
     linewidth = x$args$linewidth, line.col = x$args$line.col,
     intercept = x$args$intercept, density.col = x$args$density.col,
     density.linewidth = x$args$density.linewidth,
     density.linetype = x$args$density.linetype,
     point.size = x$args$point.size, point.linewidth = x$args$point.linewidth,
     point.linetype = x$args$point.linetype,
     point.shape = x$args$point.shape, point.col = x$args$point.col,
     ci.col = x$args$ci.col, ci.linewidth = x$args$ci.linewidth,
     ci.linetype = x$args$ci.linetype, errorbar.width = x$args$errorbar.width,
     dodge.width = x$args$dodge.width, jitter.size = x$args$jitter.size,
     jitter.width = x$args$jitter.width, jitter.height = x$args$jitter.height,
     jitter.alpha = x$args$jitter.alpha, gray = x$args$gray,
     start = x$args$start, end = x$args$end, color = x$args$color,
     xlab = x$args$xlab, ylab = x$args$ylab,
     xlim = x$args$xlim, ylim = x$args$ylim,
     xbreaks = x$args$xbreaks, ybreaks = x$args$ybreaks,
     axis.title.size = x$args$axis.title.sizes,
     axis.text.size = x$args$axis.text.size,
     strip.text.size = x$args$strip.text.size, title = x$args$title,
     subtitle = x$args$subtitle, group.col = x$args$group.col,
     plot.margin = x$args$plot.margin, legend.title = x$args$legend.title,
     legend.position = x$args$legend.position,
     legend.box.margin = x$args$legend.box.margin,
     legend.key.size = x$args$legend.key.size,
     legend.text.size = x$args$legend.text.size,
     facet.ncol = x$args$facet.ncol, facet.nrow = x$args$facet.nrow,
     facet.scales = x$args$facet.scales, filename = x$args$filename,
     width = x$args$width, height = x$args$height, units = x$args$units,
     dpi = x$args$dpi, check = TRUE, ...)
}
\arguments{
\item{x}{\code{misty.object} object.}

\item{plot}{see 'Arguments' in the functions \code{ci.*} (e.g.,
\code{\link{ci.cor}}).}

\item{bar}{see 'Arguments' in the functions \code{\link{aov.b}},
\code{\link{test.t}}, or \code{\link{test.welch}}.}

\item{box}{see 'Arguments' in the function \code{\link{test.levene}}.}

\item{violin}{see 'Arguments' in the function \code{\link{test.levene}}.}

\item{hist}{see 'Arguments' in the functions \code{ci.*} (e.g., \code{\link{ci.cor}}).}

\item{point}{see 'Arguments' in the functions \code{\link{aov.b}},
\code{ci.*} (e.g., \code{\link{ci.cor}}), or \code{\link{test.welch}}.}

\item{line}{see 'Arguments' in the functions \code{\link{aov.w}},
\code{ci.*} (e.g., \code{\link{ci.cor}}), or \code{\link{test.t}}.}

\item{ci}{see 'Arguments' in the functions \code{\link{aov.b}},
\code{\link{aov.w}}, \code{ci.*} (e.g., \code{\link{ci.cor}}),
\code{\link{test.t}}, or \code{\link{test.welch}}.}

\item{conf.level}{see 'Arguments' in the functions \code{\link{aov.b}},
\code{\link{aov.w}}, \code{ci.*} (e.g., \code{\link{ci.cor}}),
\code{\link{test.t}}, or \code{\link{test.welch}}.}

\item{adjust}{see 'Arguments' in the functions \code{\link{aov.b}},
\code{\link{aov.w}}, \code{\link{test.t}}, or
\code{\link{test.welch}},}

\item{jitter}{see 'Arguments' in the functions \code{\link{aov.b}},
\code{\link{aov.w}},  \code{\link{test.t}},
\code{\link{test.levene}},or \code{\link{test.welch}}.}

\item{density}{see 'Arguments' in the functions \code{ci.*} (e.g., \code{\link{ci.cor}}).}

\item{square}{see 'Arguments' in the function \code{\link{na.pattern}}.}

\item{rotate}{see 'Arguments' in the function \code{\link{na.pattern}}.}

\item{binwidth}{see 'Arguments' in the functions \code{ci.*} (e.g., \code{\link{ci.cor}}).}

\item{bins}{see 'Arguments' in the functions \code{ci.*} (e.g., \code{\link{ci.cor}}).}

\item{fill}{see 'Arguments' in the functions \code{ci.*} (e.g., \code{\link{ci.cor}}).}

\item{hist.alpha}{see 'Arguments' in the functions \code{ci.*} (e.g., \code{\link{ci.cor}}).}

\item{tile.alpha}{see 'Arguments' in the function \code{\link{na.pattern}}.}

\item{violin.alpha}{see 'Arguments' in the function \code{\link{test.levene}}.}

\item{violin.trim}{see 'Arguments' in the function \code{\link{test.levene}}.}

\item{box.width}{see 'Arguments' in the function \code{\link{test.levene}}.}

\item{box.alpha}{see 'Arguments' in the function \code{\link{test.levene}}.}

\item{linetype}{see 'Arguments' in the functions \code{ci.*} (e.g., \code{\link{ci.cor}})
or \code{\link{test.t}}.}

\item{linewidth}{see 'Arguments' in the function \code{\link{test.t}}.}

\item{line.col}{see 'Arguments' in the functions \code{ci.*} (e.g., \code{\link{ci.cor}}).}

\item{intercept}{see 'Arguments' in the functions \code{ci.*} (e.g., \code{\link{ci.cor}}).}

\item{density.col}{see 'Arguments' in the functions \code{ci.*} (e.g., \code{\link{ci.cor}}).}

\item{density.linewidth}{see 'Arguments' in the functions \code{ci.*} (e.g., \code{\link{ci.cor}}).}

\item{density.linetype}{see 'Arguments' in the functions \code{ci.*} (e.g., \code{\link{ci.cor}}).}

\item{point.size}{see 'Arguments' in the functions \code{\link{aov.b}},
\code{\link{aov.w}}, \code{ci.*} (e.g., \code{\link{ci.cor}}), \code{\link{test.t}},
or \code{\link{test.welch}}.}

\item{point.linewidth}{see 'Arguments' in the functions \code{ci.*} (e.g., \code{\link{ci.cor}}).}

\item{point.linetype}{see 'Arguments' in the functions \code{ci.*} (e.g., \code{\link{ci.cor}}).}

\item{point.shape}{see 'Arguments' in the functions \code{ci.*} (e.g., \code{\link{ci.cor}}).}

\item{point.col}{see 'Arguments' in the functions \code{ci.*} (e.g., \code{\link{ci.cor}}).}

\item{ci.col}{see 'Arguments' in the functions \code{ci.*} (e.g., \code{\link{ci.cor}}).}

\item{ci.linewidth}{see 'Arguments' in the functions \code{ci.*} (e.g., \code{\link{ci.cor}}).}

\item{ci.linetype}{see 'Arguments' in the functions \code{ci.*} (e.g., \code{\link{ci.cor}}).}

\item{errorbar.width}{see 'Arguments' in the functions \code{\link{aov.b}},
\code{\link{aov.w}}, \code{ci.*} (e.g., \code{\link{ci.cor}}), \code{\link{test.t}},
or \code{\link{test.welch}},}

\item{dodge.width}{see 'Arguments' in the functions \code{ci.*} (e.g., \code{\link{ci.cor}}).}

\item{jitter.size}{see 'Arguments' in the functions \code{\link{aov.b}},
\code{\link{aov.w}}, \code{\link{test.levene}},
\code{\link{test.t}}, or \code{\link{test.welch}}.}

\item{jitter.width}{see 'Arguments' in the functions \code{\link{aov.b}},
\code{\link{aov.w}}, \code{\link{test.levene}},
\code{\link{test.t}}, or \code{\link{test.welch}}.}

\item{jitter.height}{see 'Arguments' in the functions \code{\link{aov.b}},
\code{\link{test.levene}}, \code{\link{test.t}}, or
\code{\link{test.welch}}.}

\item{jitter.alpha}{see 'Arguments' in the functions \code{\link{aov.b}},
\code{\link{aov.w}}, \code{\link{test.levene}},
\code{\link{test.t}}, or \code{\link{test.welch}}.}

\item{gray}{see 'Arguments' in the functions \code{\link{multilevel.r2}}
or \code{\link{test.levene}}.}

\item{start}{see 'Arguments' in the functions \code{\link{multilevel.r2}}
or \code{\link{test.levene}}.}

\item{end}{see 'Arguments' in the functions \code{\link{multilevel.r2}}
or \code{\link{test.levene}}.}

\item{color}{see 'Arguments' in the functions \code{\link{multilevel.r2}},
\code{\link{na.pattern}}, or \code{\link{test.levene}}.}

\item{xlab}{see 'Arguments' in the functions \code{\link{aov.b}},
\code{\link{aov.w}}, \code{ci.*} (e.g., \code{\link{ci.cor}}),
\code{\link{test.levene}}, \code{\link{test.t}}, or \code{\link{test.welch}}.}

\item{ylab}{see 'Arguments' in the functions \code{\link{aov.b}},
\code{\link{aov.w}}, \code{ci.*} (e.g., \code{\link{ci.cor}}),
\code{\link{test.levene}}, \code{\link{test.t}}, or \code{\link{test.welch}}.}

\item{xlim}{see 'Arguments' in the functions \code{ci.*} (e.g., \code{\link{ci.cor}}),
\code{\link{test.levene}}, \code{\link{test.t}}, or
\code{\link{test.welch}}.}

\item{ylim}{see 'Arguments' in the functions \code{\link{aov.b}},
\code{\link{aov.w}}, \code{ci.*} (e.g., \code{\link{ci.cor}}),
\code{\link{test.levene}}, \code{\link{test.t}}, or \code{\link{test.welch}}.}

\item{xbreaks}{see 'Arguments' in the functions \code{ci.*} (e.g., \code{\link{ci.cor}}),
\code{\link{test.levene}}, \code{\link{test.t}}, or
\code{\link{test.welch}}.}

\item{ybreaks}{see 'Arguments' in the functions \code{\link{aov.b}},
\code{\link{aov.w}}, \code{ci.*} (e.g., \code{\link{ci.cor}}),
\code{\link{test.levene}}, \code{\link{test.t}}, or \code{\link{test.welch}}.}

\item{axis.title.size}{see 'Arguments' in the functions \code{ci.*} (e.g., \code{\link{ci.cor}}).}

\item{axis.text.size}{see 'Arguments' in the functions \code{ci.*} (e.g., \code{\link{ci.cor}}).}

\item{strip.text.size}{see 'Arguments' in the functions \code{ci.*} (e.g., \code{\link{ci.cor}}).}

\item{title}{see 'Arguments' in the functions \code{\link{aov.b}},
\code{\link{aov.w}}, \code{ci.*} (e.g., \code{\link{ci.cor}}),
\code{\link{test.levene}}, \code{\link{test.t}}, or \code{\link{test.welch}}.}

\item{subtitle}{see 'Arguments' in the functions \code{\link{aov.b}},
\code{\link{aov.w}}, \code{ci.*} (e.g., \code{\link{ci.cor}}),
\code{\link{test.levene}}, \code{\link{test.t}}, or \code{\link{test.welch}}.}

\item{group.col}{see 'Arguments' in the functions \code{ci.*} (e.g., \code{\link{ci.cor}}).}

\item{plot.margin}{see 'Arguments' in the functions \code{ci.*} (e.g., \code{\link{ci.cor}}).}

\item{legend.title}{see 'Arguments' in the functions \code{ci.*} (e.g., \code{\link{ci.cor}}).}

\item{legend.position}{see 'Arguments' in the functions \code{ci.*} (e.g., \code{\link{ci.cor}}).}

\item{legend.box.margin}{see 'Arguments' in the functions \code{ci.*} (e.g., \code{\link{ci.cor}})
or \code{\link{na.pattern}}.}

\item{legend.key.size}{see 'Arguments' in the function \code{\link{na.pattern}}.}

\item{legend.text.size}{see 'Arguments' in the functions \code{\link{na.pattern}}.}

\item{facet.ncol}{see 'Arguments' in the functions \code{ci.*} (e.g., \code{\link{ci.cor}}).}

\item{facet.nrow}{see 'Arguments' in the functions \code{ci.*} (e.g., \code{\link{ci.cor}}).}

\item{facet.scales}{see 'Arguments' in the functions \code{ci.*} (e.g., \code{\link{ci.cor}}).}

\item{filename}{see 'Arguments' in the functions \code{\link{aov.b}},
\code{\link{aov.w}}, \code{ci.*} (e.g., \code{\link{ci.cor}}),
\code{\link{test.levene}}, \code{\link{test.t}}, or \code{\link{test.welch}}.}

\item{width}{see 'Arguments' in the functions \code{\link{aov.b}},
\code{\link{aov.w}}, \code{ci.*} (e.g., \code{\link{ci.cor}}),
\code{\link{test.levene}}, \code{\link{test.t}}, or \code{\link{test.welch}}.}

\item{height}{see 'Arguments' in the functions \code{\link{aov.b}},
\code{\link{aov.w}}, \code{ci.*} (e.g., \code{\link{ci.cor}}),
\code{\link{test.levene}}, \code{\link{test.t}}, or \code{\link{test.welch}}.}

\item{units}{see 'Arguments' in the functions \code{\link{aov.b}},
\code{\link{aov.w}}, \code{ci.*} (e.g., \code{\link{ci.cor}}),
\code{\link{test.levene}}, \code{\link{test.t}}, or \code{\link{test.welch}}.}

\item{dpi}{see 'Arguments' in the functions \code{\link{aov.b}},
\code{\link{aov.w}}, \code{ci.*} (e.g., \code{\link{ci.cor}}),
\code{\link{test.levene}}, \code{\link{test.t}}, or \code{\link{test.welch}}.}

\item{check}{logical: if \code{TRUE} (default), argument specification
is checked.}

\item{...}{further arguments passed to or from other methods.}

}
\description{
This function plots an \code{misty.object} object.
}
\author{
Takuya Yanagida \email{takuya.yanagida@univie.ac.at}
}
