% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.result.R
\name{write.result}
\alias{write.result}
\title{Write Results of a misty Object into an Excel file}
\usage{
write.result(x, file = "Results.xlsx", tri = x$args$tri,
             digits = x$args$digits, p.digits = x$args$p.digits,
             icc.digits = x$args$icc.digits, check = TRUE)
}
\arguments{
\item{x}{misty object (\code{misty.object}) resulting from a misty function
supported by the \code{write.result} function (see 'Details').}

\item{file}{a character string naming a file with or without file extension
'.xlsx', e.g., \code{"Results.xlsx"} or \code{"Results"}.}

\item{tri}{a character string or character vector indicating which triangular
of the matrix to show on the console, i.e., \code{both} for upper and lower
triangular, \code{lower} for the lower triangular, and \code{upper} for the upper
triangular.}

\item{digits}{an integer value indicating the number of decimal places digits
to be used for displaying results.}

\item{p.digits}{an integer indicating the number of decimal places to be used
for displaying \emph{p}-values.}

\item{icc.digits}{an integer indicating the number of decimal places to be used
for displaying intraclass correlation coefficients.}

\item{check}{logical: if \code{TRUE} (default), argument specification is checked.}
}
\description{
This function writes the results of a misty object (\code{misty.object})
into a Excel file.
}
\details{
Currently the function supports result objects from the function
\code{cor.matrix}, \code{crosstab}, \code{descript}, \code{dominance.manual},
\code{dominance}, \code{effsize}, \code{freq}, \code{item.alpha}, \code{item.cfa},
\code{item.invar}, \code{item.omega}, \code{result.lca}, \code{multilevel.cfa},
\code{multilevel.cor}, \code{multilevel.descript}, \code{multilevel.fit},
\code{multilevel.invar}, \code{multilevel.omega}, \code{na.coverage},
\code{na.descript}, \code{na.pattern}, \code{robust.coef}, and \code{std.coef}.
}
\examples{
\dontrun{
#----------------------------------------------------------------------------
# Example 1: item.cfa() function

# Load data set "HolzingerSwineford1939" in the lavaan package
data("HolzingerSwineford1939", package = "lavaan")

result <- item.cfa(HolzingerSwineford1939[, c("x1", "x2", "x3")], output = FALSE)
write.result(result, "CFA.xlsx")

#----------------------------------------------------------------------------
# Example 2: multilevel.descript() function

# Load data set "Demo.twolevel" in the lavaan package
data("Demo.twolevel", package = "lavaan")

result <- multilevel.descript(y1:y3, data = Demo.twolevel, cluster = "cluster",
                              output = FALSE)
write.result(result, "Multilevel_Descript.xlsx")
}
}
\seealso{
\code{\link{cor.matrix}}, \code{\link{crosstab}}, \code{\link{descript}},
\code{\link{dominance.manual}}, \code{\link{dominance}}, \code{\link{effsize}},
\code{\link{freq}}, \code{\link{item.alpha}}, \code{\link{item.cfa}}, \code{\link{item.invar}},
\code{\link{item.omega}}, \code{\link{result.lca}}, \code{\link{multilevel.cfa}},
\code{\link{multilevel.cor}}, \code{\link{multilevel.descript}}, \code{\link{multilevel.fit}},
\code{\link{multilevel.invar}}, \code{\link{multilevel.omega}}, \code{\link{na.coverage}},
\code{\link{na.descript}}, \code{\link{na.pattern}}, \code{\link{robust.coef}}, \code{\link{std.coef}}
}
\author{
Takuya Yanagida \email{takuya.yanagida@univie.ac.at}
}
